refactor: decompose ClipBlock.tsx into focused sub-components
Closes #1026
